Move inactive tokens calculation to its own hook
